Fixed: Corrected misuse of React Hooks in `MapComponent`.
The `MapComponent` was using `useState` and `useEffect` inside a `map` function, which is a violation of the rules of Hooks. This was causing the home page to be empty.

This commit fixes the issue by extracting the marker rendering logic into its own component, `AttractionMarker`. This new component encapsulates the state and effects for a single marker, ensuring that Hooks are called at the top level of a React function component.
